refactor(terrain): the third subject solar.py was carrying
solar.py was read as two subjects, a shared solar service and the photovoltaic
product. It is three. DEM_COLLECTION, pixel_size_m, horn_slope_aspect and
fetch_dem are terrain, and terra/terrain/ already existed to hold them.

fetch_dem becomes dem.fetch_file, beside fetch and fetch_set: the three differ
by what they return, arrays over a buffered window, one ProductRead per
product, and the path of a GeoTIFF, which is what the solar terrain chain
consumes. DEM_COLLECTION dissolves into COLLECTIONS["cop30"].collection -- the
same string and the same asset key, so one literal instead of two for one
collection.

pixel_size_m and horn_slope_aspect become terra/terrain/slope.py.

TWO PIXEL SIZES IN ONE PACKAGE. slope.pixel_size_m and hand.pixel_size_m
compute the same quantity and disagree: 111_320 m per degree on both axes here
against 110_540 on the meridional one there, 0.7 percent in dy. They are not
unified by this commit, because unifying them changes the slope every solar
terrain run reports and a move is not the place to change a published number.
The divergence is stated in slope.py, and what remains is to decide which
constant is right and change the numbers on purpose.

THE ALIAS IS terrain_slope, NOT slope. Four of the rewired bodies bind a local
named slope, and `slope, _aspect = slope.horn_slope_aspect(...)` runs only
because the right-hand side is evaluated first; the next read of that name in
the same body is the array. Written the first way it also pulled the import
into two functions that never needed it, because a local `slope.shape` matched.

WHAT THIS UNLOCKS. solar.py now imports nothing from terra at all. It held
`from terra.terrain import dem` and `from terra import stac` at module top, and
carrying either into terra/sun/ would have been an immediate sibling violation
of the layers contract, since terra.terrain and terra.sun are the same layer.

Verification: pytest 502 passed, 8 skipped. gofmt, go vet and go test -race
clean. A ping through the entry point answers. The guard added in the previous
commit checks all seven rewired attribute reads.
